Refactor: Split class Map
